feat: implemented usecases for store and security routes

+++ src/main/java/com/letraaletra/api/features/user/domain/wallet/Balance.java
@@ -1,7 +1,7 @@
 package com.letraaletra.api.features.user.domain.wallet;
 
 public record Balance(
-        int coins,
-        int gems
+        long coins,
+        long gems
 ) {
 }
